Add new blank scoreboard page
