+++ app/controllers/scoreboards_controller.rb
@@ -1,8 +1,14 @@
 class ScoreboardsController < ApplicationController
   def show
-    authorize! :read, ongoing_match
-    @javascript_entrypoint = "scoreboard"
-    render layout: "minimal"
+    if (table_id = params[:id])
+      authorize! :read, Table.find(table_id)
+      @javascript_entrypoint = "scoreboard"
+      render layout: "minimal"
+    else
+      authorize! :read, ongoing_match.try!(:table)
+      @javascript_entrypoint = "legacy_scoreboard"
+      render :legacy_show, layout: "minimal"
+    end
   end
 
   def edit
