+++ app/views/scoreboards/legacy_show.html.haml
