+++ app/views/scoreboards/show.html.erb
@@ -1,0 +1,1 @@
+<div id="scoreboard"></div>
