+++ config/initializers/assets.rb
@@ -11,4 +11,5 @@ Rails.application.config.assets.version = '1.0'
 Rails.application.config.assets.precompile += %w(
   default.js
   scoreboard.js
+  legacy_scoreboard.js
 )
