+++ spec/controllers/scoreboards_controller_spec.rb
@@ -2,8 +2,31 @@ require 'rails_helper'
 
 RSpec.describe ScoreboardsController, type: :controller do
   describe "GET show" do
-    subject { get :show }
-    it { is_expected.to have_http_status :success }
+    subject { get :show, params: params }
+
+    context "without a table id" do
+      let(:params) { {} }
+      it { is_expected.to have_http_status :success }
+      it { is_expected.to render_template :legacy_show }
+    end
+
+    context "with a table id" do
+      let(:params) { { id: table_id } }
+
+      context "when the table does exist" do
+        let(:table_id) { create(:table).id }
+        it { is_expected.to have_http_status :ok }
+        it { is_expected.to render_template :show }
+      end
+
+      context "when the table doesn't exist" do
+        let(:table_id) { 2277 }
+        it "raises an error" do
+          expect { subject }.
+            to raise_error ActiveRecord::RecordNotFound
+        end
+      end
+    end
   end
 
   describe "GET edit" do
